- Removed Company model, seed and related relations - Added default Project and Event seed data - Cascade delete events from projects - Project edit/update route + modal settings - Event relation fix - Only show project list on front end - Overall fixes

+++ app/Http/Controllers/EventController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Http\Request;
-use App\Models\Company;
 use App\Models\Event;
 use App\Models\User;
 use Carbon\Carbon;
@@ -15,118 +14,113 @@ use Auth;
 
 class EventController extends Controller
 {   
-	public function show($eventid){
-
-        // $event = Event::where('id', '=', $eventid)
-        //         ->with('user')
-        //         ->first();
-
-        // return view(
-        //     'events.view',
-        //     compact('event')
-        // );
-    }
+	// public function show($eventid){
+    //     //
+    // }
     
     public function store(Request $request)	{
+
+        if($user = Auth::user()) {
+            $event = new Event;
+
+            $input = $request->input();
+
+            // Single fields
+            $event->user_id = (int)$input['user_id'];
+            $event->title = $input['title'];
+            $event->description = $input['description'];
+            $event->start = $input['start'];
+            $event->duration = ($input['duration'] ? $this->hoursToMinutes($input['duration']) : null);
+            $event->allDay = $input['allday'];
+            $event->project_id = $input['project_id'];
+            $event->price = $input['price'];
+
+            $rrule = $input['rrule'];
+            $freq = $rrule['freq'];
+
+            // Recurrence fields
+            if($freq != "ONCE") {
+
+                $rruleData = [];
+                
+                $rruleData['FREQ'] = $freq;
+                $rruleData['DTSTART'] = $input['start'];
+                
+                switch ($freq) {
+
+                    // Daily //
+                    case 'DAILY': 
+
+                        $dailyfreq = (int)$rrule['daily']['freq'];
+
+                        if($dailyfreq == 1) {
+                            $rruleData['INTERVAL'] = 1;
+                        } else if($dailyfreq == 2) {
+                            $rruleData['INTERVAL'] = (int)$rrule['daily']['interval'];
+                        }
+                        break;
+                    
+                    // Weekly //
+                    case 'WEEKLY': 
+
+                        $rruleData['INTERVAL'] = (int)$rrule['weekly']['interval'];                    
+                        $rruleData['BYDAY'] = $rrule['weekly']['bywdaylist'];
+                        break;
+
+                    // Monthly //
+                    case 'MONTHLY': 
+
+                        $monthlyfreq = $rrule['monthly']['freq'];
+                        
+                        if($monthlyfreq == 'datenum'){
+
+                            $rruleData['BYMONTHDAY'] = (int)$rrule['monthly']['datenum']['daynum'];
+                            $rruleData['INTERVAL'] = (int)$rrule['monthly']['datenum']['interval'];
+
+                        } else if($monthlyfreq == 'datename'){
         
-        $event = new Event;
+                            $rruleData['BYDAY'] = $rrule['monthly']['datename']['nth'] . $rrule['monthly']['datename']['day'];
+                            $rruleData['INTERVAL'] = (int)$rrule['monthly']['datename']['interval'];
+                        }
+                        break;
 
-        $input = $request->input();
+                    // Yearly //
+                    case 'YEARLY': 
 
-        // Single fields
-        $event->user_id = (int)$input['user_id'];
-        $event->title = $input['title'];
-        $event->description = $input['description'];
-        $event->start = $input['start'];
-        $event->duration = ($input['duration'] ? $this->hoursToMinutes($input['duration']) : null);
-        $event->allDay = $input['allday'];
-        $event->project_id = $input['project_id'];
-        $event->price = $input['price'];
-
-        $rrule = $input['rrule'];
-        $freq = $rrule['freq'];
-
-        // Recurrence fields
-        if($freq != "ONCE") {
-
-            $rruleData = [];
-            
-            $rruleData['FREQ'] = $freq;
-            $rruleData['DTSTART'] = $input['start'];
-            
-            switch ($freq) {
-
-                // Daily //
-                case 'DAILY': 
-
-                    $dailyfreq = (int)$rrule['daily']['freq'];
-
-                    if($dailyfreq == 1) {
-                        $rruleData['INTERVAL'] = 1;
-                    } else if($dailyfreq == 2) {
-                        $rruleData['INTERVAL'] = (int)$rrule['daily']['interval'];
-                    }
-                    break;
-                
-                // Weekly //
-                case 'WEEKLY': 
-
-                    $rruleData['INTERVAL'] = (int)$rrule['weekly']['interval'];                    
-                    $rruleData['BYDAY'] = $rrule['weekly']['bywdaylist'];
-                    break;
-
-                // Monthly //
-                case 'MONTHLY': 
-
-                    $monthlyfreq = $rrule['monthly']['freq'];
-                    
-                    if($monthlyfreq == 'datenum'){
-
-                        $rruleData['BYMONTHDAY'] = (int)$rrule['monthly']['datenum']['daynum'];
-                        $rruleData['INTERVAL'] = (int)$rrule['monthly']['datenum']['interval'];
-
-                    } else if($monthlyfreq == 'datename'){
-    
-                        $rruleData['BYDAY'] = $rrule['monthly']['datename']['nth'] . $rrule['monthly']['datename']['day'];
-                        $rruleData['INTERVAL'] = (int)$rrule['monthly']['datename']['interval'];
-                    }
-                    break;
-
-                // Yearly //
-                case 'YEARLY': 
-
-                    $rruleData['INTERVAL'] = (int)$rrule['yearly']['interval'];
-                    
-                    $yearlyfreq = $rrule['yearly']['freq'];
-                    
-                    if($yearlyfreq == 'date'){
-
-                        $rruleData['BYMONTHDAY'] = (int)$rrule['yearly']['date']['daynum'];
-                        $rruleData['BYMONTH'] = (int)$rrule['yearly']['date']['month'];
-
-                    } else if($yearlyfreq == 'nthdate') {
+                        $rruleData['INTERVAL'] = (int)$rrule['yearly']['interval'];
                         
-                        $rruleData['BYSETPOS'] = $rrule['yearly']['nthdate']['nth'];
-                        $rruleData['BYDAY'] = $rrule['yearly']['nthdate']['day'];
-                        $rruleData['BYMONTH'] = (int)$rrule['yearly']['nthdate']['month'];
+                        $yearlyfreq = $rrule['yearly']['freq'];
                         
-                    }                    
-                    break;
+                        if($yearlyfreq == 'date'){
+
+                            $rruleData['BYMONTHDAY'] = (int)$rrule['yearly']['date']['daynum'];
+                            $rruleData['BYMONTH'] = (int)$rrule['yearly']['date']['month'];
+
+                        } else if($yearlyfreq == 'nthdate') {
+                            
+                            $rruleData['BYSETPOS'] = $rrule['yearly']['nthdate']['nth'];
+                            $rruleData['BYDAY'] = $rrule['yearly']['nthdate']['day'];
+                            $rruleData['BYMONTH'] = (int)$rrule['yearly']['nthdate']['month'];
+                            
+                        }                    
+                        break;
+                }
+
+                // End Type
+                if($rrule['end']['endtype'] == "occurrences") {
+                    $rruleData['COUNT'] = $rrule['end']['occurrences'];
+                } else if($rrule['end']['endtype'] == "date") {
+                    $rruleData['UNTIL'] = $rrule['end']['enddate'];
+                }
+                    
+                $rrule = new RRule($rruleData);
+        
+                $event->rrule = $rrule;
             }
 
-            // End Type
-            if($rrule['end']['endtype'] == "occurrences") {
-                $rruleData['COUNT'] = $rrule['end']['occurrences'];
-            } else if($rrule['end']['endtype'] == "date") {
-                $rruleData['UNTIL'] = $rrule['end']['enddate'];
-            }
-                
-            $rrule = new RRule($rruleData);
-    
-            $event->rrule = $rrule;
+            $event->save();
+
         }
-
-        $event->save();
 
     }
 
@@ -137,9 +131,13 @@ class EventController extends Controller
 
     public function destroy($eventid){
 
-        $event = Event::find($eventid);
+        if($user = Auth::user()) {
+
+            $event = Event::find($eventid);
+            
+            $event->delete();
         
-        $event->delete();
+        }
     }
 
     public function getEventList(Request $request){
@@ -154,7 +152,7 @@ class EventController extends Controller
             'start', 
             'allday')
             ->where('user_id', $userid)
-            ->orderByDesc('updated_at')
+            ->with('project')
             ->get();
             
         foreach($events as $event) {
@@ -169,8 +167,6 @@ class EventController extends Controller
 
     public function getCalendarEvents(Request $request){
         
-        // $userid = Auth::id();
-        
         $from = new Carbon($request->from);
         $till = new Carbon($request->till);
 
@@ -183,15 +179,14 @@ class EventController extends Controller
             'duration', 
             'allDay', 
             'description')
-            // ->where('user_id', $userid)
             ->where('rrule', null)
+            ->whereHas('project')
             ->whereBetween('start', [$from, $till])
             ->orderBy('start', 'DESC')
             ->get()
             ->toArray();
         
         $rruleEvents = Event::select('id')
-            // ->where('user_id', $userid)
             ->where('rrule' , '!=', null)
             ->where('start', '<=', $till)
             ->get();
@@ -218,7 +213,7 @@ class EventController extends Controller
         foreach($events as $event) {
 
             // Convert tinyint to bool for Calendar
-            $event['allDay'] = ($event['allDay'] == 1 ? true : false); 
+            $event['allDay'] = ($event['allDay'] == 1 ? true : false);
 
             // Set endtime if it's not an allDay event
             if(!$event['allDay']){
@@ -243,6 +238,7 @@ class EventController extends Controller
             'allDay', 
             'description'
         ]);
+        
         $rrule = new RRule($event->rrule);
 
         $events = collect([]);
